requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.36
google-generativeai==0.8.5
python-telegram-bot==21.10
requests==2.32.3

python-dotenv
